Refactor CameraModel out MSCKF and FeatureEstimator
The camera model was used to project and transform pixel coordinates to
image coordinates. However if this process is brought forward towards
the FeatureTracker (i.e. the front-end), the MSCKF and FeatureEstimator
(i.e. backend) can assume the points are in image coordinates and
undistorted.

TLDR: It will now be the feature tracker's responsibility to undistort
and convert to the correct format.